+++ docs/source/apidoc.md
@@ -134,9 +134,9 @@ See Ref. :cite:`Czarnik_2021_Quantum` for more details on these methods.
    :members:
 ```
 ## Classical Shadows
-### Post-Processing
+### Classical Post-Processing
 ```{eval-rst}
-.. automodule:: mitiq.shadows.post_processing
+.. automodule:: mitiq.shadows.classical_postprocessing
    :members:
 ```
