+++ mitiq/shadows/test/test_classical_postprocessing.py
@@ -1,7 +1,7 @@
 import cirq
 import numpy as np
 
-from mitiq.shadows.post_processing import (
+from mitiq.shadows.classical_postprocessing import (
     snapshot_state,
     shadow_state_reconstruction,
     expectation_estimation_shadow,
